改组以及错误修复

+++ TaskMgrCore/handlehlp.h
@@ -2,7 +2,7 @@
 #include "stdafx.h"
 #include "ntdef.h"
 
-typedef void(*EHCALLBACK)(VOID* handle, LPWSTR type, LPWSTR name, LPWSTR address, LPWSTR objaddr, int refcount, int typeindex);
+typedef void(*EHCALLBACK)(VOID* handle, LPWSTR type, LPWSTR name, LPWSTR address, LPWSTR objaddr, ULONG refcount, int typeindex);
 
 M_CAPI(NTSTATUS) M_EH_DuplicateHandleFromProcess(HANDLE ProcessHandle, PHANDLE Handle, HANDLE HandleValue, ACCESS_MASK DesiredAccess);
 M_CAPI(BOOL) M_EH_CloseHandle(DWORD pid, LPVOID handleValue);
